Bugfixes.
Zftp does not exist any more, disable it. I've also not used it in
years, as ftp essentially is dead in usage. The Alt-m char causes
problems on some machines during zshrc generation and I've never, ever
used this binding, so just removing it makes most sense. As for
incremental it's spelled with an n and I can't believe I've left
that as is for so long.