Small fix for Anterak install script and helper for Android dev
The Anterak install script wourd not account for machines that
do not cache their xkm files. The fix was easy enough.
Android dev means I want more possible locations of the SDK.
Not super happy with my impl of that, but oh well.
The b alias is necessary because the mmm function suckily
relies on the broken sh_word_split behavior. Thankfully zsh
has local_options which saves the ass of every shell user in
the world in this situation.