Make the layout more useful.
A long-needed upgrade. Importantly, there are no more
pre-composed circumflex and diaeresis on the keyboard because
those are of rare enough use that I can't remember when they
are when I need them and I fumble around. Using dead keys is
the new way of typing them and is expected to be actually
faster (due to no more fumbling). Dead_circumflex and
dead_diaeresis moved to a more premium space.

- Make dead_circumflex and dead_diaeresis the first two dead
  keys on the first row. Shift the rest to the right.
- Move copyright to a better place, and make trademark the
  shift-copyright symbol.
- Move up/down arrow and left/right guillemot to make room
  for more interesting things on the left/right arrow
  neighborhood.
- Introduce check mark and ballot x
- Introduce left/right arrow, and have this and the left
  and right arrow have their double version on shift.
  In particular, double right arrow was sorely needed.
- Fill the bottom row from the left with circled numerals.
  That makes up to 5, the relevance of which will be
  evaluated. Black circled numbers are on shift.
- Add bullet on AltGr-tilde and smiley on the shift of that.
- Opportunistically tab arrows and double reciprocal arrows on
  AltGr+Tab.

Also a small number of cleanups (use symbol instead of code point
number when possible, re-align)