Improvements to Anterak.
Symbols like sixsuperior, sevensuperior see no use at all in
practice. Higher circled numbers do, to some extent. Replace
that. Keep the footnotes up to 3, because that's what sees use
in practice. Keep ½, ⅓, ¼ as the most used of fractions, though
they also don't see a lot of use.

It's also more intuitive, both for ¹²³/½⅓¼ and ①②③/❶❷❸, but also
for mathematical symbols like ∃∀∄∈. Because this is net positive
on freed keys, also allows to add ≈≠ which is wanted sometimes, at
the expense of infinity that is never used.