Fix prompts again
The problem was in the quoting, not the final $ as I stated
in the cover comment of the previous patch.
Also, only display the machine and do that in bold. I don't
exactly use multiple accounts on any of my machines, save
for root which is red anyway. The name of the machine is
sufficient to know where I am.