Clean up zsh conf
These functions have seen no use in years, have no possible
use in the future, and in the extreme case where I'd want to
remember them, well that's what git is for.

Note that option multibyte was removed because it serves no
purpose, as it's on by default if the shell supports it and
can't be turned on anyway if the shell doesn't.