Fix xkb_geometry and add comments
Comments to document how this searches for sections,
as this has always been puzzling.

For a reason I don't understand there seem to be no longer
be a pc(jp106) geometry, at least on my work machine. This
would prevent installing this layout, meaning in particular
the henkan and katakana keys would no longer map to
iso_level3_shift (aka AltGr).

Really what should be done here is a new geometry file for
jp109, and ideally upstream that.