Add an install script.
Traditionally I had my conf directly as a .git in $HOME. This
is inconvenient for a number of reasons ; a big one is, my
entire $HOME becomes a git repo so commands like `git status`
will do something anywhere under my home, when it's usually
not relevant (e.g. `git status` outside of a git repo should
give an error, not give me a status for an unrelated repo,
it's confusing). Also $HOME is full of stuff not managed in
the git so `git status` would show hundreds of unrelated
files that I'd ignore every time.

Now that's pretty stupid, but I never got arount to fixing
it on my home machine, though on newer installs I've taken
to put the git is some subdir and symlinking to the git-
managed files, which is a lot more sane.

But now I find that for some tools, $HOME/.xkb is actually
special. For example, this is preventing the kwin tests
from running, because they try to find a definition for a
us keyboard with pc104 in ~/.xkb when it's not meant for
that at all. Or maybe it was a long time ago, but I'm
pretty sure it no longer is. We'll see if anything breaks...

Now with this fixed and out of the way, it's important to
remember to not just ln -s everything from this dir to
$HOME. Obviously .git, but also en_JP and its readme which
have nothing to do in $HOME, and .xkb which is not meant
for that. In fact it will be renamed in the next patch,
since it should not start with a dot.

The best and most automatic way to do this is to have a
script to manage the correct files to link, so this patch
does that. I hope I won't forget to update the script next
time I put some new dot file that I want to manage in here.