Fix the numpad
...and make it quite richer.

Note that this is not meant to be quite final ; it is meant as a first
test version, to test how useful it is. It can be made yet richer and
probably easier to use.

Also fix an include ordering issue in the geometry : naming the file
'geometry/pc' means xkbcomp will either use the local file or the
system-wide 'geometry/pc' but not merge them, making it difficult
to xkbcomp locally. The install script used to meld it manually into
the system-wide 'geometry/pc' file at quite some complexity, but
this is not only more complex than having a different file, it will
also be overwritten by any update of 'geometry/pc', so this is much
better.