Add a script that toggles screensaver inhibit
...which I bind on C-M-esc.

This relies on jdaemon being installed, because AFAIK there is
no way to do that without some persistent process having this
functionality, because any held lock is cleared when the process
holding it terminates.
Even keeping the script running wouldn't work because in this
case it's the qdbus process that would hold the lock, and *that*
terminates.